refactor to put API in server package
