+++ internal/server/api.go
@@ -1,60 +1,20 @@
-package api
+package server
 
 import (
 	"encoding/json"
-	"fmt"
-	"html"
 	"io"
 	"net/http"
-	"regexp"
 	"strconv"
 
 	// "github.com/jessesomerville/yodahunters/internal/envconfig"
 	"github.com/jessesomerville/yodahunters/internal/log"
-	"github.com/jessesomerville/yodahunters/internal/pg"
 )
 
-func HandleRequest(r *http.Request, w http.ResponseWriter) {
-	// Testing code - just reflect the POST body in a 200
-
-	// \/api\/(.+)
-	pathPattern := regexp.MustCompile(`\/api\/(.+)`)
-	pathMatch := pathPattern.FindStringSubmatch(r.URL.Path)
-
-	// There shouldn't be a way to hit this handler without a match...
-	if len(pathMatch) != 2 {
-		log.Errorf(r.Context(), "Request with URL %q hit API Handler and failed the path regex!\nPathMatch %s", r.URL.Path, pathMatch)
-		http.Error(w, "Invalid URL Path", http.StatusInternalServerError)
-		return
-	}
-
-	path := pathMatch[1]
-
-	defer r.Body.Close()
-
-	bodyBytes, err := io.ReadAll(r.Body)
-	if err != nil {
-		http.Error(w, "Failed to read request body", http.StatusInternalServerError)
-		log.Errorf(r.Context(), "Failed to read request body!")
-	}
-
-	body := string(bodyBytes)
-
-	fmt.Fprintf(w, "Received the following body from path %s\n\n%s", html.EscapeString(path), html.EscapeString(body))
-}
-
-func HandleGetThreads(w http.ResponseWriter, r *http.Request) {
+func (s *Server) HandleGetThreads(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 
-	dbClient, err := pg.NewClient(ctx, "postgres")
-	if err != nil {
-		http.Error(w, "Failed to acquire DB client", http.StatusInternalServerError)
-		log.Errorf(r.Context(), "Failed to acquire DB client")
-	}
-	defer dbClient.Close(ctx)
-
 	var threads []Thread
-	rows, err := dbClient.Query(ctx, "SELECT * FROM threads")
+	rows, err := s.dbClient.Query(ctx, "SELECT * FROM threads")
 	if err != nil {
 		http.Error(w, "Failed to query DB", http.StatusInternalServerError)
 		log.Errorf(r.Context(), "Failed to query DB")
@@ -79,7 +39,7 @@ func HandleGetThreads(w http.ResponseWriter, r *http.Request) {
 	w.Write(jsonData)
 }
 
-func HandleGetThreadByID(w http.ResponseWriter, r *http.Request) {
+func (s *Server) HandleGetThreadByID(w http.ResponseWriter, r *http.Request) {
 	id, err := strconv.Atoi(r.PathValue("id"))
 	if err != nil {
 		http.Error(w, "Error retrieving ID from Path", http.StatusInternalServerError)
@@ -88,15 +48,8 @@ func HandleGetThreadByID(w http.ResponseWriter, r *http.Request) {
 
 	ctx := r.Context()
 
-	dbClient, err := pg.NewClient(ctx, "postgres")
-	if err != nil {
-		http.Error(w, "Failed to acquire DB client", http.StatusInternalServerError)
-		log.Errorf(r.Context(), "Failed to acquire DB client")
-	}
-	defer dbClient.Close(ctx)
-
 	var thread Thread
-	err = dbClient.QueryRow(ctx, "SELECT id, title, body FROM threads WHERE id = $1", id).Scan(&thread.ID, &thread.Title, &thread.Body)
+	err = s.dbClient.QueryRow(ctx, "SELECT id, title, body FROM threads WHERE id = $1", id).Scan(&thread.ID, &thread.Title, &thread.Body)
 	if err != nil {
 		http.Error(w, "Failed to retrieve thread!", http.StatusInternalServerError)
 		log.Errorf(ctx, "Failed to retrieve thread with ID: %d", id)
@@ -112,7 +65,7 @@ func HandleGetThreadByID(w http.ResponseWriter, r *http.Request) {
 
 }
 
-func HandlePostThreads(w http.ResponseWriter, r *http.Request) {
+func (s *Server) HandlePostThreads(w http.ResponseWriter, r *http.Request) {
 	const insertThreadQuery = `
 	INSERT INTO threads (title, body)
 	VALUES ($1, $2)
@@ -135,15 +88,9 @@ func HandlePostThreads(w http.ResponseWriter, r *http.Request) {
 	}
 
 	ctx := r.Context()
-	dbClient, err := pg.NewClient(ctx, "postgres")
-	if err != nil {
-		http.Error(w, "Failed to acquire DB client", http.StatusInternalServerError)
-		log.Errorf(r.Context(), "Failed to acquire DB client")
-	}
-	defer dbClient.Close(ctx)
 
 	var thread Thread
-	err = dbClient.QueryRow(ctx, insertThreadQuery, t.Title, t.Body).Scan(&thread.ID, &thread.Title, &thread.Body)
+	err = s.dbClient.QueryRow(ctx, insertThreadQuery, t.Title, t.Body).Scan(&thread.ID, &thread.Title, &thread.Body)
 	if err != nil {
 		http.Error(w, "Failed to update threads table", http.StatusInternalServerError)
 		log.Errorf(ctx, "Couldn't update threads table!")
