+++ internal/server/models.go
@@ -1,4 +1,4 @@
-package api
+package server
 
 // TODO
 type Thread struct {
